fix: tests triggering different failures than they supposed
This tests basically tests that the state machine expires and fails,
but the underlying condition that is causing it, is it's own doing
(gw client thinks it didn't join the federation), and not what was
supposed to be an error (expired invoice).

In other words without this change the error in the logs is:

```
2024-04-18T22:39:03.018509Z ERROR pay_invoice{payload=PayInvoicePayload { federation_id: FederationId(24a34a182503fc0729a90cebe7759247cdf1c4143d0200daddab5ff174a8ace3), ... ln_gateway::rpc::rpc_server: error=Invalid Metadata: No federation with id 24a34a182503fc0729a90cebe7759247cdf1c4143d0200daddab5ff174a8ace3
```

With it:

```
Outgoing Payment Error OutgoingContractError: An invalid contract was specified.
```