chore(ci): start cross-compile jobs earlier
This was introduced to limit amount of concurrency, and would be fine
if not for the fact that in the merge queue we do need to run
cross-compilation on MacOS which is slow, and our MQ is running
serially. Because of that we want to start cross-compilation
jobs sooner than later, so by the time other tests are done
cross-compilation on MacOS is done as well.